perf(ai): rewrite AI tab as a fast, minimal Codex-style workspace
The previous AI tab was a single 2.6k-line Insights.tsx. Typing lagged
because the composer read textarea.scrollHeight on every keystroke (a
forced synchronous reflow of the whole conversation), and the tab pulled
heavy projections on mount (today-timeline rebuild + recap range). New
chat blocked on a createThread round-trip.

Rewrite it as small, individually-memoized pieces under
src/renderer/views/insights so typing, streaming, and searching never
re-render each other:

- AICompose: isolated local input state + CSS `field-sizing: content`
  auto-grow. Zero per-keystroke layout reads. Forwards a focus() handle.
- MessageList: memoized; re-renders only on real message-list events,
  not on streaming chunks (those flow through the streaming store).
- HistorySearch: the local-history search kept verbatim in behaviour
  (debounced search.all with highlighted excerpts), restyled as a slim
  top bar whose results drop into an overlay so they never push the chat.
- useAIChat: owns chat state + actions. Lighter mount — no recap range,
  no today-timeline rebuild; CLI detection only runs when a CLI provider
  is selected. New chat is optimistic (server auto-creates the thread on
  first send), so it's instant.
- AIWorkspace: the shell — top bar (search + new chat + history),
  centered empty hero with starter prompts, docked composer.

Preserved: streaming + output sanitization, threads (list/create/load/
delete/deep-link), copy/rate/retry, follow-up chips, focus action cards,
inline artifact cards, ConnectAI, analytics events. The global Cmd/Ctrl+K
command palette in App.tsx is unchanged.

Removed non-chat surfaces from the tab: recap panel, hero summary
generator, and the dedicated Files browser/preview pane (artifacts still
appear as inline cards in their message; artifact deep links open the
file directly).

Insights.tsx is now a one-line re-export of AIWorkspace, keeping the
route and the focusMetricCopy guard. rendererHookSafety guards updated
to the new module layout (search keeps its own state; composer does no
scrollHeight read).
