+++ src/renderer/styles/globals.css
@@ -310,3 +310,41 @@ a[href], button.link-style {
   80%  { transform: scale(0.96); }
   100% { transform: scale(1); }
 }
+
+/* ── AI tab (Codex-minimal) ─────────────────────────────────────────────────
+   The composer auto-grows to its content with pure CSS (field-sizing).
+   This removes the per-keystroke scrollHeight read — a forced synchronous
+   reflow against the whole conversation tree — that made typing feel frozen. */
+.ai-composer-input {
+  field-sizing: content;
+}
+
+@keyframes aiMessageIn {
+  from { opacity: 0; transform: translateY(6px); }
+  to   { opacity: 1; transform: translateY(0); }
+}
+
+.ai-message-in {
+  animation: aiMessageIn 180ms cubic-bezier(0.22, 0.61, 0.36, 1) both;
+}
+
+@keyframes aiCaretBlink {
+  0%, 45% { opacity: 1; }
+  55%, 100% { opacity: 0.15; }
+}
+
+.ai-caret {
+  display: inline-block;
+  width: 7px;
+  height: 14px;
+  margin-left: 2px;
+  vertical-align: text-bottom;
+  border-radius: 2px;
+  background: var(--color-primary);
+  animation: aiCaretBlink 1s steps(1) infinite;
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .ai-message-in { animation: none; }
+  .ai-caret { animation: none; }
+}
