+++ src/renderer/views/insights/AICompose.tsx
@@ -1,45 +1,46 @@
-import { memo, useEffect, useRef, useState } from 'react'
+import type { ForwardedRef } from 'react'
+import { forwardRef, memo, useImperativeHandle, useRef, useState } from 'react'
+import { IconSend } from './icons'
+
+export interface AIComposeHandle {
+  focus: () => void
+}
 
 interface AIComposeProps {
   onSubmit: (text: string) => void
   loading: boolean
+  placeholder?: string
 }
 
-function IconSend() {
-  return (
-    <svg width="14" height="14" viewBox="0 0 16 16" fill="none" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-      <path d="M3 13 13 3" />
-      <path d="M5.5 3H13v7.5" />
-    </svg>
-  )
-}
-
-function AIComposeImpl({ onSubmit, loading }: AIComposeProps) {
+// Isolated, self-contained composer.
+//
+// Why this is fast:
+//   1. Input text lives in this component's local state, so a keystroke
+//      re-renders ONLY this component — never the message list, the search
+//      bar, or the workspace shell (all siblings, all memoized).
+//   2. The textarea auto-grows with CSS `field-sizing: content`
+//      (`.ai-composer-input` in globals.css). The old composer read the
+//      textarea's scroll-height on every keystroke inside a rAF, which forces a
+//      synchronous layout of the whole conversation tree — that is what made
+//      typing feel frozen. There is no layout measurement read here at all.
+function AIComposeImpl(
+  { onSubmit, loading, placeholder }: AIComposeProps,
+  ref: ForwardedRef<AIComposeHandle>,
+) {
   const [input, setInput] = useState('')
   const textareaRef = useRef<HTMLTextAreaElement>(null)
 
-  // Auto-resize the textarea to fit content (1–7 visual lines).
-  // Read scrollHeight exactly once and reuse it: this effect runs on every
-  // keystroke, and reading a layout property forces a synchronous reflow. The
-  // AI tab mounts a long conversation tree, so a second forced reflow per
-  // keystroke is what makes typing feel like it freezes.
-  useEffect(() => {
-    const frame = window.requestAnimationFrame(() => {
-      const textarea = textareaRef.current
-      if (!textarea) return
-      textarea.style.height = 'auto'
-      const contentHeight = textarea.scrollHeight
-      textarea.style.height = `${Math.min(Math.max(contentHeight, 24), 140)}px`
-      textarea.style.overflowY = contentHeight > 140 ? 'auto' : 'hidden'
-    })
-    return () => window.cancelAnimationFrame(frame)
-  }, [input])
+  useImperativeHandle(ref, () => ({
+    focus: () => textareaRef.current?.focus(),
+  }), [])
 
   const send = () => {
     const text = input.trim()
     if (!text || loading) return
     onSubmit(text)
     setInput('')
+    // Keep focus on the composer so the next message can be typed immediately.
+    textareaRef.current?.focus()
   }
 
   const trimmed = input.trim()
@@ -47,16 +48,17 @@ function AIComposeImpl({ onSubmit, loading }: AIComposeProps) {
   return (
     <div style={{
       display: 'flex',
-      alignItems: 'center',
-      gap: 10,
-      borderRadius: 18,
+      alignItems: 'flex-end',
+      gap: 8,
+      borderRadius: 16,
       border: '1px solid var(--color-border-ghost)',
       background: 'var(--color-surface)',
-      padding: '10px 10px 10px 16px',
+      padding: '8px 8px 8px 16px',
       boxShadow: 'var(--color-shadow-floating)',
     }}>
       <textarea
         ref={textareaRef}
+        className="ai-composer-input"
         value={input}
         onChange={(event) => setInput(event.target.value)}
         onKeyDown={(event) => {
@@ -67,20 +69,22 @@ function AIComposeImpl({ onSubmit, loading }: AIComposeProps) {
         }}
         disabled={loading}
         rows={1}
+        autoFocus
         aria-label="Ask Daylens about your work history"
-        placeholder="Ask about your day, or ask for a report, chart, table, or export..."
+        placeholder={placeholder ?? 'Ask anything about your work…'}
         style={{
           flex: 1,
-          minHeight: 20,
-          maxHeight: 140,
+          minHeight: 24,
+          maxHeight: 184,
+          overflowY: 'auto',
           border: 'none',
           background: 'transparent',
           outline: 'none',
           color: 'var(--color-text-primary)',
           fontSize: 13.5,
-          lineHeight: '20px',
+          lineHeight: '22px',
           resize: 'none',
-          padding: '8px 0',
+          padding: '6px 0',
           display: 'block',
         }}
       />
@@ -90,8 +94,8 @@ function AIComposeImpl({ onSubmit, loading }: AIComposeProps) {
         type="button"
         aria-label="Send message"
         style={{
-          width: 36,
-          height: 36,
+          width: 34,
+          height: 34,
           padding: 0,
           borderRadius: 999,
           border: 'none',
@@ -102,6 +106,7 @@ function AIComposeImpl({ onSubmit, loading }: AIComposeProps) {
           alignItems: 'center',
           justifyContent: 'center',
           flexShrink: 0,
+          transition: 'background 160ms ease, color 160ms ease',
         }}
       >
         <IconSend />
@@ -111,7 +116,6 @@ function AIComposeImpl({ onSubmit, loading }: AIComposeProps) {
 }
 
 // React.memo so chunk-driven parent re-renders skip the composer entirely.
-// Default shallow comparison is sufficient: onSubmit is wrapped in a
-// useRef-stable callback by the parent, and `loading` is the only prop the
-// composer actually wants to react to.
-export const AICompose = memo(AIComposeImpl)
+// `onSubmit` is a useRef-stable callback from the parent, so the composer's
+// only real re-render trigger is the `loading` prop.
+export const AICompose = memo(forwardRef<AIComposeHandle, AIComposeProps>(AIComposeImpl))
