+++ src/renderer/views/insights/MessageList.tsx
@@ -1,0 +1,274 @@
+import { memo, useState } from 'react'
+import type { AIMessageAction, FocusSession } from '@shared/types'
+import { ipc } from '../../lib/ipc'
+import { MarkdownMessage } from './markdown'
+import { StreamingMessage } from './StreamingMessage'
+import {
+  IconActionButton,
+  IconArtifactFile,
+  IconCopy,
+  IconExternal,
+  IconRetry,
+  IconThumbsDown,
+  IconThumbsUp,
+} from './icons'
+import {
+  actionFeedbackKey,
+  messageActionKey,
+  type ActionFeedbackEntry,
+  type MessageActionStateEntry,
+  type ThreadMessage,
+} from './types'
+
+type ReviewFocusAction = Extract<AIMessageAction, { kind: 'review_focus_session' }>
+
+function FocusReviewActionCard({
+  action,
+  state,
+  onSave,
+}: {
+  action: ReviewFocusAction
+  state?: MessageActionStateEntry
+  onSave: (note: string) => void
+}) {
+  const [draft, setDraft] = useState(action.suggestedNote ?? '')
+  return (
+    <div style={{ borderRadius: 12, border: '1px solid var(--color-border-ghost)', background: 'var(--color-surface-low)', padding: 12, display: 'grid', gap: 8 }}>
+      <div style={{ fontSize: 12.5, color: 'var(--color-text-secondary)', lineHeight: 1.6 }}>
+        Save a short reflection to this focus session.
+      </div>
+      <textarea
+        aria-label="Focus session review"
+        value={draft}
+        onChange={(event) => setDraft(event.target.value)}
+        placeholder={action.placeholder ?? 'Add a short focus review'}
+        rows={4}
+        style={{ width: '100%', resize: 'vertical', borderRadius: 10, border: '1px solid var(--color-border-ghost)', background: 'var(--color-surface)', color: 'var(--color-text-primary)', padding: '10px 12px', fontSize: 12.5, lineHeight: 1.6, outline: 'none' }}
+      />
+      <div style={{ display: 'flex', alignItems: 'center', gap: 10, flexWrap: 'wrap' }}>
+        <button
+          type="button"
+          onClick={() => onSave(draft)}
+          disabled={state?.busy}
+          style={{ padding: '8px 12px', borderRadius: 9, border: '1px solid var(--color-border-ghost)', background: 'var(--color-surface)', color: 'var(--color-text-primary)', fontSize: 12.5, fontWeight: 700, cursor: state?.busy ? 'default' : 'pointer', opacity: state?.busy ? 0.7 : 1 }}
+        >
+          {state?.busy ? 'Saving…' : action.label}
+        </button>
+        {state?.successLabel && <span style={{ fontSize: 12, color: 'var(--color-focus-green)' }}>{state.successLabel}</span>}
+        {state?.error && <span style={{ fontSize: 12, color: '#f87171' }}>{state.error}</span>}
+      </div>
+    </div>
+  )
+}
+
+export interface MessageListProps {
+  messages: ThreadMessage[]
+  latestCompletedAssistantId: string | number | undefined
+  actionFeedback: Record<string, ActionFeedbackEntry>
+  messageActionState: Record<string, MessageActionStateEntry>
+  reducedMotion: boolean
+  activeFocusSession: FocusSession | null
+  onCopy: (messageId: string | number, content: string, answerKind: ThreadMessage['answerKind']) => void
+  onRate: (message: ThreadMessage, rating: 'up' | 'down' | null) => void
+  onRetry: (index: number, message: ThreadMessage) => void
+  onMessageAction: (messageId: string | number, action: AIMessageAction, options?: { reviewNote?: string }) => void
+  onFollowUpClick: (message: ThreadMessage, suggestionText: string, source: string) => void
+  scrollToBottom: () => void
+}
+
+function MessageListImpl({
+  messages,
+  latestCompletedAssistantId,
+  actionFeedback,
+  messageActionState,
+  reducedMotion,
+  activeFocusSession,
+  onCopy,
+  onRate,
+  onRetry,
+  onMessageAction,
+  onFollowUpClick,
+  scrollToBottom,
+}: MessageListProps) {
+  return (
+    <div style={{ display: 'grid', gap: 24, contain: 'layout' }}>
+      {messages.map((message, index) => (
+        message.role === 'user' ? (
+          <div key={message.id} className="ai-message-in" style={{ display: 'flex', justifyContent: 'flex-end' }}>
+            <div style={{ maxWidth: '76%', borderRadius: '14px 14px 6px 14px', background: 'var(--color-accent-dim)', color: 'var(--color-primary)', padding: '11px 14px', fontSize: 13, fontWeight: 550, whiteSpace: 'pre-wrap' }}>
+              {message.content}
+            </div>
+          </div>
+        ) : (
+          <div key={message.id} className="ai-message-in" style={{ display: 'flex' }}>
+            <div style={{
+              flex: 1,
+              maxWidth: 720,
+              lineHeight: 1.6,
+              ...(message.state === 'error' ? {
+                borderRadius: 12,
+                border: '1px solid rgba(248, 113, 113, 0.28)',
+                background: 'rgba(248, 113, 113, 0.08)',
+                padding: '14px 16px 10px',
+              } : {}),
+            }}>
+              {message.state === 'pending' ? (
+                <StreamingMessage
+                  messageId={String(message.id)}
+                  fallback={<div style={{ fontSize: 13, color: 'var(--color-text-tertiary)' }}>Thinking<span className="ai-caret" /></div>}
+                  renderContent={(text) => <><MarkdownMessage content={text} /><span className="ai-caret" /></>}
+                  onSnapshotUpdate={scrollToBottom}
+                />
+              ) : (
+                <>
+                  {message.state === 'error' && (
+                    <div style={{ fontSize: 11.5, fontWeight: 700, letterSpacing: '0.08em', textTransform: 'uppercase', color: '#f87171', marginBottom: 8 }}>
+                      Provider error
+                    </div>
+                  )}
+                  <MarkdownMessage content={message.content} />
+
+                  {(message.actions?.length ?? 0) > 0 && (
+                    <div style={{ display: 'grid', gap: 10, marginTop: 14 }}>
+                      {message.actions?.map((action) => {
+                        const key = messageActionKey(message.id, action)
+                        const state = messageActionState[key]
+                        if (action.kind === 'review_focus_session') {
+                          return (
+                            <FocusReviewActionCard
+                              key={key}
+                              action={action}
+                              state={state}
+                              onSave={(reviewNote) => onMessageAction(message.id, action, { reviewNote })}
+                            />
+                          )
+                        }
+                        const disabled = state?.busy
+                          || (action.kind === 'start_focus_session' && Boolean(activeFocusSession))
+                          || (action.kind === 'stop_focus_session' && activeFocusSession?.id !== action.sessionId)
+                        const contextHint = action.kind === 'start_focus_session' && activeFocusSession
+                          ? 'A focus session is already active.'
+                          : action.kind === 'stop_focus_session' && activeFocusSession?.id !== action.sessionId
+                            ? 'That focus session is no longer active.'
+                            : null
+                        return (
+                          <div key={key} style={{ borderRadius: 12, border: '1px solid var(--color-border-ghost)', background: 'var(--color-surface-low)', padding: 12, display: 'flex', alignItems: 'center', justifyContent: 'space-between', gap: 12, flexWrap: 'wrap' }}>
+                            <div style={{ fontSize: 12.5, color: 'var(--color-text-secondary)', lineHeight: 1.6 }}>
+                              {contextHint ?? (action.kind === 'start_focus_session' ? 'Start a focus session from this chat context.' : 'Stop the active focus session from here.')}
+                            </div>
+                            <div style={{ display: 'flex', alignItems: 'center', gap: 10, flexWrap: 'wrap' }}>
+                              <button
+                                type="button"
+                                onClick={() => onMessageAction(message.id, action)}
+                                disabled={disabled}
+                                style={{ padding: '8px 12px', borderRadius: 9, border: '1px solid var(--color-border-ghost)', background: 'var(--color-surface)', color: 'var(--color-text-primary)', fontSize: 12.5, fontWeight: 700, cursor: disabled ? 'default' : 'pointer', opacity: disabled ? 0.7 : 1 }}
+                              >
+                                {state?.busy ? (action.kind === 'start_focus_session' ? 'Starting…' : 'Stopping…') : action.label}
+                              </button>
+                              {state?.successLabel && <span style={{ fontSize: 12, color: 'var(--color-focus-green)' }}>{state.successLabel}</span>}
+                              {state?.error && <span style={{ fontSize: 12, color: '#f87171' }}>{state.error}</span>}
+                            </div>
+                          </div>
+                        )
+                      })}
+                    </div>
+                  )}
+
+                  {(message.artifacts?.length ?? 0) > 0 && (
+                    <div style={{ display: 'grid', gap: 10, marginTop: 14 }}>
+                      {message.artifacts?.map((artifact) => {
+                        const color = artifact.format === 'csv' ? '#16a34a' : artifact.format === 'html' ? '#7c3aed' : artifact.format === 'json' ? '#f59e0b' : '#2563eb'
+                        const kind = artifact.format === 'csv' ? 'csv' : artifact.format === 'html' ? 'html_chart' : 'markdown'
+                        return (
+                          <button
+                            key={`${message.id}:${artifact.id}`}
+                            type="button"
+                            onClick={() => void ipc.shell.openPath(artifact.path)}
+                            style={{ display: 'flex', alignItems: 'center', gap: 10, padding: '10px 12px', borderRadius: 10, border: '1px solid var(--color-border-ghost)', background: 'var(--color-surface)', cursor: 'pointer', textAlign: 'left', width: '100%' }}
+                          >
+                            <div style={{ width: 34, height: 34, borderRadius: 8, background: `${color}15`, border: `1px solid ${color}28`, color, display: 'flex', alignItems: 'center', justifyContent: 'center', flexShrink: 0 }}>
+                              <IconArtifactFile kind={kind} />
+                            </div>
+                            <div style={{ flex: 1, minWidth: 0 }}>
+                              <div style={{ fontSize: 12.5, fontWeight: 600, color: 'var(--color-text-primary)', overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap' }}>
+                                {artifact.title}
+                              </div>
+                              <div style={{ fontSize: 11, color: 'var(--color-text-tertiary)', marginTop: 2 }}>
+                                {artifact.format.toUpperCase()} · click to open
+                              </div>
+                            </div>
+                            <span style={{ color: 'var(--color-text-tertiary)', flexShrink: 0, display: 'inline-flex' }}><IconExternal /></span>
+                          </button>
+                        )
+                      })}
+                    </div>
+                  )}
+
+                  {message.id === latestCompletedAssistantId && message.state === 'complete' && (message.suggestedFollowUps?.length ?? 0) >= 2 && (
+                    <div style={{ display: 'flex', flexWrap: 'wrap', gap: 8, marginTop: 14 }}>
+                      {message.suggestedFollowUps?.map((suggestion) => (
+                        <button
+                          key={`${message.id}:${suggestion.text}`}
+                          type="button"
+                          onClick={() => onFollowUpClick(message, suggestion.text, suggestion.source)}
+                          style={{ padding: '7px 12px', borderRadius: 999, border: '1px solid var(--color-border-ghost)', background: 'transparent', color: 'var(--color-text-secondary)', cursor: 'pointer', fontSize: 12.5 }}
+                        >
+                          {suggestion.text}
+                        </button>
+                      ))}
+                    </div>
+                  )}
+
+                  <div style={{ display: 'flex', alignItems: 'center', gap: 8, flexWrap: 'wrap', marginTop: 12 }}>
+                    <IconActionButton
+                      label="Copy response"
+                      feedbackLabel={actionFeedback[actionFeedbackKey(message.id, 'copy')]?.success ? 'Copied' : undefined}
+                      success={actionFeedback[actionFeedbackKey(message.id, 'copy')]?.success ?? false}
+                      pulseNonce={actionFeedback[actionFeedbackKey(message.id, 'copy')]?.pulseNonce ?? 0}
+                      reducedMotion={reducedMotion}
+                      onClick={() => onCopy(message.id, message.content, message.answerKind)}
+                    >
+                      <IconCopy />
+                    </IconActionButton>
+                    <IconActionButton
+                      label="Thumbs up"
+                      tone="positive"
+                      selected={message.rating === 'up'}
+                      pulseNonce={actionFeedback[actionFeedbackKey(message.id, 'up')]?.pulseNonce ?? 0}
+                      reducedMotion={reducedMotion}
+                      onClick={() => onRate(message, message.rating === 'up' ? null : 'up')}
+                    >
+                      <IconThumbsUp />
+                    </IconActionButton>
+                    <IconActionButton
+                      label="Thumbs down"
+                      tone="negative"
+                      selected={message.rating === 'down'}
+                      pulseNonce={actionFeedback[actionFeedbackKey(message.id, 'down')]?.pulseNonce ?? 0}
+                      reducedMotion={reducedMotion}
+                      onClick={() => onRate(message, message.rating === 'down' ? null : 'down')}
+                    >
+                      <IconThumbsDown />
+                    </IconActionButton>
+                    {message.id === latestCompletedAssistantId && (
+                      <IconActionButton
+                        label="Retry response"
+                        pulseNonce={actionFeedback[actionFeedbackKey(message.id, 'retry')]?.pulseNonce ?? 0}
+                        reducedMotion={reducedMotion}
+                        onClick={() => onRetry(index, message)}
+                      >
+                        <IconRetry />
+                      </IconActionButton>
+                    )}
+                  </div>
+                </>
+              )}
+            </div>
+          </div>
+        )
+      ))}
+    </div>
+  )
+}
+
+export const MessageList = memo(MessageListImpl)
